Add initial table
