+++ epic-document-api/docker-compose.yml
@@ -11,7 +11,7 @@ services:
       - POSTGRES_DB=epic-document
       - POSTGRES_HOST_AUTH_METHOD=trust
     ports:
-      - 54332:5432/tcp
+      - 54334:5432/tcp
     restart: unless-stopped
 
   epic-document-api-db-test:
@@ -24,7 +24,7 @@ services:
       - POSTGRES_DB=epic-document
       - POSTGRES_HOST_AUTH_METHOD=trust
     ports:
-      - 54333:5432/tcp
+      - 54335:5432/tcp
     restart: unless-stopped
 
   keycloak:
