+++ epic-document-api/src/epic_document_api/logging.conf
