+++ epic-document-api/src/epic_document_api/models/base_model.py
