+++ epic-document-api/src/epic_document_api/utils/run_version.py
@@ -11,8 +11,19 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+"""Supply version and commit hash info."""
+import os
 
-"""This exports all of the models and schemas used by the application."""
+from epic_document_api.version import __version__
 
-from .db import db, ma, migrate
-from .user import User
+
+def _get_build_openshift_commit_hash():
+    return os.getenv('OPENSHIFT_BUILD_COMMIT', None)
+
+
+def get_run_version():
+    """Return a formatted version string for this service."""
+    commit_hash = _get_build_openshift_commit_hash()
+    if commit_hash:
+        return f'{__version__}-{commit_hash}'
+    return __version__
