--- epic-document-api/src/scaffold_api/services/user_service.py
@@ -1,40 +1,0 @@
-"""Service for user management."""
-from scaffold_api.models.user import User as UserModel
-
-
-class UserService:
-    """User management service."""
-
-    @classmethod
-    def get_user_by_id(cls, _user_id):
-        """Get user by id."""
-        db_user = UserModel.find_by_id(_user_id)
-        return db_user
-
-    @classmethod
-    def get_all_users(cls):
-        """Get all users."""
-        users = UserModel.get_all()
-        return users
-
-    @classmethod
-    def create_user(cls, user_data):
-        """Create user."""
-        created_user = UserModel.create_user(user_data)
-        return created_user
-
-    @classmethod
-    def update_user(cls, user_id, user_data):
-        """Update user."""
-        updated_user = UserModel.update_user(user_id, user_data)
-        return updated_user
-
-    @classmethod
-    def delete_user(cls, user_id):
-        """Update user."""
-        user = UserModel.find_by_id(user_id)
-        if not user:
-            return None
-
-        user.delete()
-        return user
